Description: fix the problems about wearable devices compile alarms IssueNo: https://gitee.com/openharmony/graphic_ui/issues/I53JQ8 Feature or Bugfix: Bugfix Binary Source:No Signed-off-by: liuyuxiang <liuyuxiang7@huawei.com>

+++ frameworks/src/core/animation/transition_impl.cpp
@@ -313,10 +313,10 @@ void TransitionImpl::PerformTransition(int16_t from,
             break;
         case TransitionType::OPACITY: {
             double rate = (double)updateAttrValue / ALPHA_MAX;
-            view_->SetStyle(STYLE_BACKGROUND_OPA, viewStatus_.rectOpacity * rate);
-            view_->SetStyle(STYLE_IMAGE_OPA, viewStatus_.imageOpacity * rate);
-            view_->SetStyle(STYLE_LINE_OPA, viewStatus_.lineOpacity * rate);
-            view_->SetStyle(STYLE_TEXT_OPA, viewStatus_.textOpacity * rate);
+            view_->SetStyle(STYLE_BACKGROUND_OPA, static_cast<int64_t>(viewStatus_.rectOpacity * rate));
+            view_->SetStyle(STYLE_IMAGE_OPA, static_cast<int64_t>(viewStatus_.imageOpacity * rate));
+            view_->SetStyle(STYLE_LINE_OPA, static_cast<int64_t>(viewStatus_.lineOpacity * rate));
+            view_->SetStyle(STYLE_TEXT_OPA, static_cast<int64_t>(viewStatus_.textOpacity * rate));
             break;
         }
         default:
@@ -327,7 +327,7 @@ void TransitionImpl::PerformTransition(int16_t from,
     invalidatedArea.Join(invalidatedArea, view_->GetRect());
     view_->InvalidateRect(invalidatedArea);
 }
-int8_t TransitionImpl::GetNumIterations(const char* iterations)
+int8_t TransitionImpl::GetNumIterations(const char *iterations)
 {
     int8_t min = 1;
     int8_t max = 127;
@@ -345,7 +345,7 @@ int8_t TransitionImpl::GetNumIterations(const char* iterations)
     return (int8_t)value;
 }
 
-bool TransitionImpl::IsEndWith(const char* src, const char* end)
+bool TransitionImpl::IsEndWith(const char *src, const char *end)
 {
     if ((src == nullptr) || (end == nullptr)) {
         return false;
@@ -385,9 +385,9 @@ void TransitionImpl::ResetRepeatParam()
     if (animator_ != nullptr) {
         animator_->SetRunTime(0); // reset animation executing time
     }
-    timeArrivaled_ = false;   // reset timeArrivaled flag
-    bgcolorTimeSrc_ = 0;      // reset bg-color last update time
-    count_ = 1;               // reset bg-color update number of times
+    timeArrivaled_ = false; // reset timeArrivaled flag
+    bgcolorTimeSrc_ = 0;    // reset bg-color last update time
+    count_ = 1;             // reset bg-color update number of times
 }
 
 void TransitionImpl::RecordViewStatus()
@@ -410,7 +410,7 @@ void TransitionImpl::RecoveryViewStatus(Rect invalidatedAreaBefore) const
     view_->SetY(viewStatus_.y);
     view_->SetHeight(viewStatus_.height);
     view_->SetWidth(viewStatus_.width);
-    TransformMap& transMap = view_->GetTransformMap();
+    TransformMap &transMap = view_->GetTransformMap();
     Polygon polygon(Rect(0, 0, 0, 0));
     transMap.SetPolygon(polygon);
     view_->SetStyle(STYLE_BACKGROUND_COLOR, viewStatus_.background_color.full);
